+++ vitrina/smart_contracts/views.py
@@ -83,26 +83,87 @@ class BaseAgreementMixin:
         )
 
 
-class AgreementListView(
-    LoginRequiredMixin,
-    BaseProjectMixin,
-    PermissionRequiredMixin,
-    TemplateView,
-):
+class BaseAgreementListView(LoginRequiredMixin, PermissionRequiredMixin, TemplateView):
     model = Agreement
-    template_name = "smart_contracts/agreement_list.html"
+    template_name = None
 
-    detail_url_name = "project-detail"
-    history_url_name = "project-history"
+    paginate_by = 10
 
-    project: Project
+    parent: object
+    parent_type: str
+
+    def get_queryset(self):
+        """Override in subclasses, agreement queryset will be filtered by parent."""
+        raise NotImplementedError
 
     def has_permission(self) -> bool:
-        return can_view_agreements(self.request.user, self.project)
+        """Override in subclasses, permission depends on parent."""
+        return False
+
+    def get_context_data(self, **kwargs: Any):
+        context = super().get_context_data(**kwargs)
+
+        agreements = self.get_queryset()
+        paginator = Paginator(agreements, self.paginate_by)
+        page = paginator.get_page(self.request.GET.get("page"))
+
+        context.update(
+            {
+                "agreements": agreements,
+                "agreement_status_descriptions": AGREEMENT_STATUS_DESCRIPTIONS,
+                "page_obj": page,
+                "paginator": paginator,
+                "can_create_agreements": False,  # Override in subclasses.
+                "parent": self.parent,
+                "parent_type": self.parent_type,
+            }
+        )
+
+        return context
+
+
+class BaseAgreementDetailView(LoginRequiredMixin, PermissionRequiredMixin, TemplateView):
+    template_name = None
+    parent: object = None
+    parent_type: str = None
+
+    agreement: Agreement
+
+    def has_permission(self) -> bool:
+        """Override in subclasses."""
+        return False
+
+    def get_context_data(self, **kwargs: Any) -> dict:
+        context = super().get_context_data(**kwargs)
+
+        context.update(
+            {
+                "agreement": self.agreement,
+                "project": self.agreement.project,
+                "agreement_files": self.agreement.files.all().order_by("-created_at"),
+                "agreement_status_descriptions": AGREEMENT_STATUS_DESCRIPTIONS,
+                "page_title": self.agreement.detail_page_title,
+                "parent": self.parent,
+                "parent_type": self.parent_type,
+            }
+        )
+
+        return context
+
+
+class AgreementListView(BaseProjectMixin, BaseAgreementListView):
+    template_name = "smart_contracts/agreement_list.html"
+    parent_type = "project"
+
+    def setup(self, request, *args: Any, **kwargs: Any) -> None:
+        super().setup(request, *args, **kwargs)
+        self.parent: Project = self.get_project(kwargs["pk"])
+        return None
 
     def dispatch(self, request, *args, **kwargs):
         dispatch = super().dispatch(request, *args, **kwargs)
-        if not self.project.organization:
+
+        if not self.parent.organization:
             messages.error(
                 self.request,
                 _("Panaudojimo atvejis registruotas fizinio asmens vardu negali turėti sutarčių."),
@@ -110,48 +171,26 @@ class AgreementListView(
             return HttpResponseRedirect(reverse("project-detail", kwargs={"pk": self.project.pk}))
         return dispatch
 
-    def get_context_data(self, **kwargs: Any) -> dict:
+    def has_permission(self) -> bool:
+        return can_view_agreements(self.request.user, self.project)
+
+    def get_queryset(self):
+        return get_agreements(self.request.user).filter(project=self.project)
+
+    def get_context_data(self, **kwargs: Any):
         context = super().get_context_data(**kwargs)
-
-        project_agreements = get_agreements(self.request.user).filter(project=self.project)
-
-        paginator = Paginator(project_agreements, 10)
-        page_number = self.request.GET.get("page")
-        page = paginator.get_page(page_number)
-
-        context.update(
-            {
-                "project": self.project,
-                "agreements": page.object_list,
-                "agreement_status_descriptions": AGREEMENT_STATUS_DESCRIPTIONS,
-                "page_obj": page,
-                "paginator": paginator,
-                "can_create_agreements": can_create_agreements(self.request.user, self.project),
-            }
-        )
-        context["parent_links"].update(
-            {
-                None: _("Sutartys"),
-            }
-        )
+        context["can_create_agreements"] = can_create_agreements(self.request.user, self.project)
+        context["parent_links"].update({None: _("Sutartys")})
         return context
 
 
-class AgreementDetailView(
-    LoginRequiredMixin,
-    BaseProjectMixin,
-    BaseAgreementMixin,
-    PermissionRequiredMixin,
-    TemplateView,
-):
-    model = Agreement
+class AgreementDetailView(BaseProjectMixin, BaseAgreementMixin, BaseAgreementDetailView):
     template_name = "smart_contracts/agreement_detail.html"
+    parent_type = "project"
 
-    detail_url_name = "project-detail"
-    history_url_name = "project-history"
-
-    project: Project
-    agreement: Agreement
+    def setup(self, request: WSGIRequest, *args: Any, **kwargs: Any) -> None:
+        super().setup(request, *args, **kwargs)
+        self.parent: Project = self.get_project(kwargs["pk"])
 
     def has_permission(self) -> bool:
         return can_view_agreement(self.request.user, self.agreement)
@@ -161,10 +200,6 @@ class AgreementDetailView(
 
         context.update(
             {
-                "agreement": self.agreement,
-                "agreement_files": self.agreement.files.all().order_by("-created_at"),
-                "agreement_status_descriptions": AGREEMENT_STATUS_DESCRIPTIONS,
-                "page_title": self.agreement.detail_page_title,
                 "can_create_agreements": can_create_agreements(self.request.user, self.project),
                 "can_submit_agreements": can_submit_agreements(self.request.user, self.agreement),
                 "can_approve_agreements": can_approve_agreements(self.request.user, self.agreement),
@@ -174,12 +209,9 @@ class AgreementDetailView(
                 "can_upload_agreement_file": can_upload_agreement_file(self.request.user, self.agreement),
             }
         )
-        context["parent_links"].update(
-            {
-                reverse("agreement-list", args=[self.project.pk]): _("Sutartys"),
-                None: self.agreement.detail_page_title,
-            }
-        )
+
+        context["parent_links"].update({reverse("agreement-list", args=[self.project.pk]): _("Sutartys")})
+
         return context